% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_token.R
\name{add_token}
\alias{add_token}
\title{Set Lens Access Token}
\usage{
add_token(token = NULL, overwrite = FALSE)
}
\arguments{
\item{token}{Your Lens.org API token}

\item{overwrite}{Whether to overwrite existing token}
}
\value{
Invisibly, the API key
}
\description{
This helper function allows you to set your API key in .Renviron file
}
\details{
modified from \url{https://github.com/cwthom/TargomoR/blob/master/R/setup.R}
}
\examples{
\dontrun{
# write to a global file at Sys.getenv("HOME")
addtoken(token = "YOUR_LENS_TOKEN")
}

}
\concept{token}
