% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\title{Get Collection Data from Lens}
\usage{
get_collection(collection_id, token = NULL)
}
\arguments{
\item{collection_id}{Collection ID number in Lens.org}

\item{token}{Your Lens.org API token, add to .Renviron via add_token()
function}
}
\value{
Content from Lens Scholarly Data
}
\description{
Get Collection Data from Lens
}
\seealso{
Other Get Data: 
\code{\link{get_scholarly_data}()},
\code{\link{get_scholarly_df}()}
}
\concept{Get Data}
