% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scholarly_data.R
\name{get_scholarly_data}
\alias{get_scholarly_data}
\title{Get Scholarly Data from Lens}
\usage{
get_scholarly_data(query, token = NULL)
}
\arguments{
\item{query}{Your query}

\item{token}{Your Lens.org API token, add to .Renviron via add_token()
function}
}
\value{
Content from Lens Collection
}
\description{
Get Scholarly Data from Lens
}
\seealso{
Other Get Data: 
\code{\link{get_collection}()}
}
\concept{Get Data}
