% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scholarly_df.R
\name{get_scholarly_df}
\alias{get_scholarly_df}
\title{Get Scholarly Data from Lens}
\usage{
get_scholarly_df(query, token = NULL, output = "df")
}
\arguments{
\item{query}{Your query}

\item{token}{Your Lens.org API token, add to .Renviron via add_token()
function}

\item{output}{select df (dataframe) of json, default is df}
}
\value{
Content from Lens Collection as dataframe or json
}
\description{
Get Scholarly Data from Lens
}
\seealso{
Other Get Data: 
\code{\link{get_collection}()},
\code{\link{get_scholarly_data}()}
}
\concept{Get Data}
