% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens2r-package.R
\docType{package}
\name{lens2r-package}
\alias{lens2r}
\alias{lens2r-package}
\title{lens2r: R Interface to the Lens API}
\description{
The Package is an R Interface to the Lens API <https://www.lens.org>
    Requires access token from lens.org.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbalci/lens2r}
  \item Report bugs at \url{https://github.com/sbalci/lens2r/issues}
}

}
\author{
\strong{Maintainer}: Serdar Balci \email{drserdarbalci@gmail.com} (\href{https://orcid.org/0000-0002-7852-3851}{ORCID})

}
\keyword{internal}
